for stupid resason
the reason is, when i work at in windows with lunarvim with the flutter-tools plugin, it does not reacognize the lib\main_development.dart file. Even if i use the ```FlutterRun --flavor development``` it stills doesn't find the lib\main.dart file. So, for that reason i had to create this little piece of file